requirements.txt
aiofiles
alembic
black
certifi
cffi
charset-normalizer
click
cryptography
ecdsa
fastapi
gitdb
gitpython
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httpx
idna
iniconfig
jinja2
mako
markdown
markupsafe
mypy
mypy-extensions
ollama
packaging
passlib
pathspec
platformdirs
pluggy
proto-plus
protobuf
psutil
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-core
pytest
pytest-asyncio
pytest-cov
python-jose
python-multipart
pyyaml
requests
rich
rsa
ruff
six
smmap
sniffio
sqlalchemy
starlette
typed-ast
typing-extensions
urllib3
uvicorn
uvloop
watchfiles
weasyprint
websockets
networkx
